refactor(config): extract transport-config; trim grab-bag surface (#159)
Issue #159 — `src/config.ts` was a 322-line, 40-export grab-bag
mixing five unrelated concerns AND leaking
`@huggingface/inference`'s `InferenceProviderOrPolicy` type to every
consumer of `HUGGINGFACE_PROVIDER`. Three changes here:

src/transport-config.ts (new):
- Move the entire RFC 008 transport block out of `config.ts`:
  `McpTransport`, `DEFAULT_MCP_PORT`, `DEFAULT_MCP_BIND_ADDR`,
  `TransportConfig`, `TransportConfigError`, `loadTransportConfig`,
  `normalizeOrigin`, plus their internal helpers
  `parseTransport`/`parsePort`/`parseAllowedOrigins`. ~117 lines
  moved verbatim.
- Update consumers: `KnowledgeBaseServer.ts` splits its config
  imports across `./config.js` and `./transport-config.js`;
  `transport/base-http-host.ts` and `transport/sse.test.ts` switch
  to the new module.

src/config.ts:
- Drop `huggingFaceRouterUrl` from the public surface (unexport).
  Zero external consumers; only used internally by
  `HUGGINGFACE_ENDPOINT_URL`'s default.
- Mark `parseReindexTriggerPollMs` `@internal` via JSDoc tag.
  It's exported only so `config.test.ts` can pin parser semantics
  (default, sentinel, MIN/MAX clamp, scientific-notation accept) —
  every production caller uses the resolved
  `REINDEX_TRIGGER_POLL_MS` constant.
- Re-type `HUGGINGFACE_PROVIDER` from `InferenceProviderOrPolicy`
  to plain `string`. The SDK type was leaking to every consumer
  even though the runtime value is a string. Drop the
  `@huggingface/inference` type import from `config.ts` entirely.

src/embedding-provider.ts:
- Cast `HUGGINGFACE_PROVIDER as InferenceProviderOrPolicy` at the
  SDK boundary (the only call site that needs the SDK shape).
- Add the matching `import type { InferenceProviderOrPolicy }`
  here, where the HuggingFace SDK is the legitimate concern.

Sizes:
  src/config.ts: 322 → 219 (-103 lines)
  src/transport-config.ts: +122 (new)
  Total `config.ts`-area surface still fits in a single grok-able
  file when you read it; transport plumbing is now its own seam.

Behaviour preserved:
- Every transport-config error wording is unchanged.
- The pre-#159 `HUGGINGFACE_PROVIDER`-typed call site behaves
  identically; only the module-boundary type tightens.
- `parseReindexTriggerPollMs` stays exported (the JSDoc tag is
  advisory; tests still import it directly).

530/530 tests pass — same baseline as the prior PR.

Closes #159.
